コンポーネントに分割
