+++ frontend/src/app/components/menu/DominationMenu.module.css
@@ -1,0 +1,54 @@
+.governanceFrame {
+    position: fixed;
+    top: 64px;
+    right: 27px;
+    width: 250px;
+    height: 350px;
+    padding: 0px 15px 47px 15px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+  }
+
+  .menuFrameBackground {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background: url('/frame/menu.png') no-repeat center center;
+    background-size: 100% 100%;
+    z-index: 2;
+  }
+
+  .governanceInner {
+    width: 95%;
+    height: 95%;
+    background-color: #FFFFFF;
+    padding: 20px 10px 30px 10px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: flex-start;
+    border-radius: 5px;
+    position: relative;
+    z-index: 1;
+  }
+  
+  .governanceTable {
+    width: 100%;
+    border-spacing: 0 14px;
+    font-size: 25px;
+    margin-top: -14px;
+    line-height: 1;
+  }
+
+  .menuLine {
+    width: 250px;
+    height: 20px;
+    background: url('/frame/line.png') no-repeat center center;
+    background-size: 100% auto;
+    position: relative;
+    z-index: 1;
+  }
